ref(preprod): Rename file to preprod_artifact_snapshot_latest_base
