+++ src/sentry/preprod/api/endpoints/preprod_artifact_snapshot_latest_base.py
